+++ src/main/kotlin/com/atlassian/performance/tools/jiraperformancetests/api/RegressionResults.kt
@@ -1,8 +1,7 @@
 package com.atlassian.performance.tools.jiraperformancetests.api
 
-import com.atlassian.performance.tools.report.api.result.CohortResult
-
+@Deprecated("Give us feedback about your use case.")
 class RegressionResults(
-    val baseline: CohortResult,
-    val experiment: CohortResult
+    @Suppress("DEPRECATION") val baseline: com.atlassian.performance.tools.report.api.result.CohortResult,
+    @Suppress("DEPRECATION") val experiment: com.atlassian.performance.tools.report.api.result.CohortResult
 )